feat: 임베딩 코드 함수화
